Revert "修改潜能点数据类型"
This reverts commit 5f582055cdf7e8e7abd26c6a829c8f85db253724.

# Conflicts:
#	gs/beans/fire/pb/RoleDetail.java
#	gs/src/fire/pb/GameSystemConfig.java
#	编译Java代码工具/gsxdb.jar

+++ gs/src/fire/pb/effect/RoleImpl.java
@@ -242,11 +242,11 @@ public class RoleImpl extends AttrCalcRole
 			return null;
 		if(fire.pb.buff.Module.existState(roleid, fire.pb.buff.BuffConstant.StateType.STATE_BATTLE_FIGHTER))
 			return null;
-		final long curpoint = prop.getPoint().get(prop.getScheme());
+		final int curpoint = prop.getPoint().get(prop.getScheme());
 		final int addpoints = agi + cons + endu + iq + str;
 		if (addpoints > curpoint)
 			return null;
-		prop.getPoint().put(prop.getScheme(), (long)(curpoint - addpoints));		
+		prop.getPoint().put(prop.getScheme(), (curpoint - addpoints));		
 		RoleAddPointProperties addfp = prop.getAddpointfp();
 		
 		xbean.BasicFightProperties bfp = prop.getBfp();
